fix: instrumentation.node.ts を復元してwebpackビルドエラーを解消
- instrumentation.ts を元の動的インポートパターンに戻す
  （@grpc/grpc-js がwebpackビルドに巻き込まれる問題を解消）
- instrumentation.node.ts を復元（Node.js専用モジュール隔離層）
- tsconfig の moduleResolution を node16 から node10 に修正
  （module: commonjs との不整合エラーを解消）
- next-env.d.ts を本番ビルド形式に更新
