+++ apps/api-ai/tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
         "module": "commonjs",
-        "moduleResolution": "node",
+        "moduleResolution": "node10",
         "target": "es2020",
         "outDir": "dist",
         "rootDir": ".",
