+++ apps/web/instrumentation.ts
@@ -1,62 +1,15 @@
 /**
  * Next.js Instrumentation Hook
  *
- * Application Insights v3 SDK を useAzureMonitor() API で初期化する。
+ * Node.js ランタイム専用の処理は instrumentation.node.ts に分離している。
+ * これにより Edge ランタイム向け webpack コンパイル時に
+ * @grpc/grpc-js 等の Node.js 固有モジュールが解析されるのを防ぐ。
  *
- * 認証方式: AAD認証（マネージドID）
- * - APPI リソース `appi-pm-exam-dx` は DisableLocalAuth=true のため、
- *   Instrumentation Key / 接続文字列だけでは 401 になる。
- * - ManagedIdentityCredential を使用してテレメトリを送信する。
- *
- * 接続文字列: TELEMETRY_CONNECTION_STRING（カスタム名）から読み取る。
- * Linux App Service の IPA コードレスエージェントは APPLICATIONINSIGHTS_* や
- * APPINSIGHTS_* プレフィックスの環境変数を検出して自動有効化し、
- * 手動 SDK の OpenTelemetry セットアップと競合するため、
- * IPA が認識しない環境変数名を使用している。
+ * 参照: https://nextjs.org/docs/app/building-your-application/optimizing/instrumentation
  */
 export async function register() {
-    const runtime = process.env.NEXT_RUNTIME;
-    if (runtime === 'edge') return;
-    if (typeof window !== 'undefined') return;
-
-    const connectionString = process.env.TELEMETRY_CONNECTION_STRING;
-    if (!connectionString) {
-        console.warn('[AppInsights] TELEMETRY_CONNECTION_STRING not set, telemetry disabled');
-        return;
-    }
-
-    try {
-        // クラウドロール名を OpenTelemetry 環境変数で設定（useAzureMonitor より前に必要）
-        if (!process.env.OTEL_SERVICE_NAME) {
-            process.env.OTEL_SERVICE_NAME = 'pm-exam-dx-web';
-        }
-
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const appInsightsModule = await import('applicationinsights') as any;
-        const appInsights = (appInsightsModule.default ?? appInsightsModule) as typeof import('applicationinsights');
-
-        // AAD認証: ManagedIdentityCredential を使用
-        // APPI リソースの DisableLocalAuth=true に対応
-        const { ManagedIdentityCredential } = await import('@azure/identity');
-        const credential = new ManagedIdentityCredential();
-
-        appInsights.useAzureMonitor({
-            azureMonitorExporterOptions: {
-                connectionString,
-                credential,
-            },
-            instrumentationOptions: {
-                http: { enabled: true },
-                console: { enabled: true },
-                azureSdk: { enabled: true },
-            } as import('applicationinsights').InstrumentationOptions,
-            enableAutoCollectExceptions: true,
-            enableAutoCollectPerformance: true,
-            enableLiveMetrics: false,
-        });
-
-        console.log('[AppInsights] SDK initialized (useAzureMonitor v3 + AAD auth)');
-    } catch (error) {
-        console.error('[AppInsights] Failed to initialize SDK:', error);
+    if (process.env.NEXT_RUNTIME === 'nodejs') {
+        const { registerNodeInstrumentation } = await import('./instrumentation.node');
+        await registerNodeInstrumentation();
     }
 }
